Dungeon object updates

+++ src/app/emu/memory/mock_memory.h
@@ -23,11 +23,11 @@ class MockClock : public Clock {
 
 class MockMemory : public Memory {
  public:
-  MOCK_CONST_METHOD1(ReadByte, uint8_t(uint16_t address));
-  MOCK_CONST_METHOD1(ReadWord, uint16_t(uint16_t address));
-  MOCK_CONST_METHOD1(ReadWordLong, uint32_t(uint16_t address));
+  MOCK_CONST_METHOD1(ReadByte, uint8_t(uint32_t address));
+  MOCK_CONST_METHOD1(ReadWord, uint16_t(uint32_t address));
+  MOCK_CONST_METHOD1(ReadWordLong, uint32_t(uint32_t address));
   MOCK_METHOD(std::vector<uint8_t>, ReadByteVector,
-              (uint16_t address, uint16_t length), (const, override));
+              (uint32_t address, uint16_t length), (const, override));
 
   MOCK_METHOD2(WriteByte, void(uint32_t address, uint8_t value));
   MOCK_METHOD2(WriteWord, void(uint32_t address, uint16_t value));
@@ -76,20 +76,20 @@ class MockMemory : public Memory {
   void Init() {
     ON_CALL(*this, ReadByte(::testing::_))
         .WillByDefault(
-            [this](uint16_t address) { return memory_.at(address); });
+            [this](uint32_t address) { return memory_.at(address); });
     ON_CALL(*this, ReadWord(::testing::_))
-        .WillByDefault([this](uint16_t address) {
+        .WillByDefault([this](uint32_t address) {
           return static_cast<uint16_t>(memory_.at(address)) |
                  (static_cast<uint16_t>(memory_.at(address + 1)) << 8);
         });
     ON_CALL(*this, ReadWordLong(::testing::_))
-        .WillByDefault([this](uint16_t address) {
+        .WillByDefault([this](uint32_t address) {
           return static_cast<uint32_t>(memory_.at(address)) |
                  (static_cast<uint32_t>(memory_.at(address + 1)) << 8) |
                  (static_cast<uint32_t>(memory_.at(address + 2)) << 16);
         });
     ON_CALL(*this, ReadByteVector(::testing::_, ::testing::_))
-        .WillByDefault([this](uint16_t address, uint16_t length) {
+        .WillByDefault([this](uint32_t address, uint16_t length) {
           std::vector<uint8_t> data;
           for (int i = 0; i < length; i++) {
             data.push_back(memory_.at(address + i));
